+++ source/extensions/lightspeed.trex.project_wizard.core/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.18]
+### Changed
+- Adding layer validation for existing project file
+
 ## [1.2.17]
 ### Changed
 - Update deps